PlayerListRefactoring
