+++ Assets/Code/PhotonLogin.cs
@@ -8,18 +8,15 @@ using UnityEngine.UI;
 
 public sealed class PhotonLogin : MonoBehaviourPunCallbacks
 {
-    [Header("Prefabs")]
-    [SerializeField] private PlayerListElementView _playerListElementPrefab;
-
+    [Header("Managers")]
+    [SerializeField] private PlayerListManager _playerListManager;
+    
     [Header("Panels")]
-    [SerializeField] private PlayerListPanelView _playerListPanelView;
     [SerializeField] private RoomAdminPanelView _roomAdminPanelView;
     [SerializeField] private RoomListPanelView _roomListPanelView;
     [SerializeField] private GameObject _inventoryPanel;
 
     [SerializeField] private Button _showRoomsButton;
-
-    private readonly List<PlayerListElementView> _playerElements = new List<PlayerListElementView>();
 
     private void Awake()
     {
@@ -27,6 +24,8 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
         _roomListPanelView.OnJoinRoomButtonClicked += JoinSelectedRoom;
         _roomListPanelView.OnCreateRoomButtonClicked += CreateRoom;
         _showRoomsButton.onClick.AddListener(OnShowRoomsButtonClicked);
+
+        _playerListManager.OnKickPlayer += KickPlayer;
     }
 
     private void OnDestroy()
@@ -34,11 +33,8 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
         _roomListPanelView.OnJoinRoomButtonClicked -= JoinSelectedRoom;
         _roomListPanelView.OnCreateRoomButtonClicked -= CreateRoom;
         _showRoomsButton.onClick.RemoveListener(OnShowRoomsButtonClicked);
-
-        foreach (var element in _playerElements)
-        {
-            element.OnKickButtonClicked -= KickPlayer;
-        }
+        
+        _playerListManager.OnKickPlayer -= KickPlayer;
     }
 
     private void Start()
@@ -102,7 +98,6 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
 
     public override void OnCreatedRoom()
     {
-        base.OnCreatedRoom();
         _roomListPanelView.gameObject.SetActive(false);
         _roomAdminPanelView.gameObject.SetActive(true);
         _roomAdminPanelView.OnPrivacyButtonClicked += SwitchRoomPrivacy;
@@ -111,17 +106,8 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
 
     public override void OnLeftRoom()
     {
-        foreach (var element in _playerElements)
-        {
-            if (element != null)
-            {
-                element.OnKickButtonClicked -= KickPlayer;
-                 Destroy(element.gameObject);
-            }
-        }
-        
-        _playerListPanelView.gameObject.SetActive(false);
-        
+        _playerListManager.OnLeftRoom();
+
         if (_roomAdminPanelView == null || !_roomAdminPanelView.gameObject.activeSelf) return;
         _roomAdminPanelView.gameObject.SetActive(false);
         _roomAdminPanelView.OnPrivacyButtonClicked -= SwitchRoomPrivacy;
@@ -138,19 +124,11 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
         Debug.LogError($"Create room failed! Code: {returnCode}, message: {message}");
     }
 
-    //TODO: Move methods below to PlayerListManager or something
     public override void OnJoinedRoom()
     {
-        foreach (var player in PhotonNetwork.PlayerList)
-        {
-            Debug.Log($"Added player {player.NickName} | {player.UserId}");
-            CreatePlayerElement(player);
-        }
+        _playerListManager.OnJoinedRoom();
         
         _inventoryPanel.SetActive(false);
-        
-        _playerListPanelView.SetRoomName(PhotonNetwork.CurrentRoom.Name);
-        _playerListPanelView.gameObject.SetActive(true);
     }
 
     private void KickPlayer(Player playerToKick)
@@ -160,27 +138,12 @@ public sealed class PhotonLogin : MonoBehaviourPunCallbacks
 
     public override void OnPlayerEnteredRoom(Player newPlayer)
     {
-        base.OnPlayerEnteredRoom(newPlayer);
-        CreatePlayerElement(newPlayer);
-    }
-
-    private void CreatePlayerElement(Player newPlayer)
-    {
-        var newElement = Instantiate(_playerListElementPrefab, _playerListPanelView.ElementsRoot);
-        newElement.gameObject.SetActive(true);
-        newElement.Initialize(newPlayer, newPlayer.NickName);
-        newElement.OnKickButtonClicked += KickPlayer;
-        _playerElements.Add(newElement);
+        _playerListManager.OnPlayerEnteredRoom(newPlayer);
     }
 
     public override void OnPlayerLeftRoom(Player otherPlayer)
     {
-        var elementToDelete = _playerElements.FirstOrDefault(element => element.PlayerActorNumber == otherPlayer.ActorNumber);
-        if (elementToDelete)
-        {
-            Destroy(elementToDelete.gameObject);
-            elementToDelete.OnKickButtonClicked -= KickPlayer;
-        }
+        _playerListManager.OnPlayerLeftRoom(otherPlayer);
     }
 
     private void OnStartGameButtonClicked()
