+++ Assets/Scenes/MainProfile.unity
@@ -727,7 +727,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &121559307
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2791,8 +2791,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d3eba37ccad54f58b6c8e7634b705893, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _playerListElementPrefab: {fileID: 4360805025317316656, guid: 893d60ba5e93e354d937d1075f2ccc99, type: 3}
-  _playerListPanelView: {fileID: 1359234334}
+  _playerListManager: {fileID: 645253436}
   _roomAdminPanelView: {fileID: 531588784}
   _roomListPanelView: {fileID: 121559310}
   _inventoryPanel: {fileID: 532050937}
@@ -3671,9 +3670,9 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 0, y: 0}
+  m_SizeDelta: {x: -17, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &480781981
 MonoBehaviour:
@@ -4730,6 +4729,7 @@ GameObject:
   - component: {fileID: 645253434}
   - component: {fileID: 645253433}
   - component: {fileID: 645253435}
+  - component: {fileID: 645253436}
   m_Layer: 0
   m_Name: Managers
   m_TagString: Untagged
@@ -4819,6 +4819,20 @@ MonoBehaviour:
   _characterLevel1: {fileID: 481257203}
   _characterName2: {fileID: 971151582}
   _characterLevel2: {fileID: 671287491}
+--- !u!114 &645253436
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 645253430}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4151fe05f7ce45f3884baf9578dd0eab, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _playerListElementPrefab: {fileID: 4360805025317316656, guid: 893d60ba5e93e354d937d1075f2ccc99, type: 3}
+  _playerListPanelView: {fileID: 1359234334}
 --- !u!1 &660947209
 GameObject:
   m_ObjectHideFlags: 0
@@ -6283,7 +6297,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 20}
   m_Pivot: {x: 0.5, y: 0.5}
