+++ UserSettings/EditorUserSettings.asset
@@ -15,10 +15,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c78111b12563f22213229
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c7d1f19030b3e3e29367e38271427fb
+      value: 22424703114646680e0b0227036c72111f192838232e2f3c2867083debf42d
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c72111f192838232e2f3c2867083debf42d
+      value: 22424703114646680e0b0227036c7d1f19030b3e3e29367e38271427fb
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
